requirements.txt
fastapi
uvicorn
openai
tiktoken
python-dotenv

PyJWT
pydantic[email]
pydantic-settings
sqlalchemy
pillow
rembg
slowapi
httpx
pytest-asyncio
boto3 
python-multipart
python-dotenv
cryptography
psycopg2-binary
python-jose==3.3.0
geopy
pyswisseph
kerykeion

torch
torchvision
scikit-image
opencv-python
pillow-heif
mediapipe